chore: cut prompt-prose pins, a barrel tautology, and source-regex tails
Reading the small files turned up two species the scans had no signal for.

Barrel tautology — sandbox-export.test.ts imported the same seven symbols
through `../index.js` and `../sandbox/index.js` and asserted they were
`===`. The module system guarantees that, and a dropped re-export is a
compile error in this very file, so the test could never fail on its own.

Prompt-prose pins — swarm-mode.test.ts and graph-mode.test.ts each render
a prompt and assert it contains ten English phrases ("preferred default
execution strategy", "Do not manufacture parallelism"). They assert
nothing about behavior and break on any rewording, which is what prompts
get. Deleted both. Prompt tests that check interpolation stay, since
those verify the values a prompt carries rather than its prose —
session-environment-prompt asserting `Git branch: main` is a real test.

Source-regex tails, trimmed rather than deleted:

- project-context-badge kept the two tests that resolve real git
  metadata in tmpdirs (normal checkout, worktree gitdir, upward root
  walk) and lost the seven that regexed main.ts for exact expressions
  like `fallbackRoots: () => [process.cwd(), app.getAppPath()]`.
  273 lines → 54.
- attachment-frontend-contract had one test that renders a ChatView with
  an image attachment; the other seven regexed source and CSS tokens.
  Reduced to that one and renamed to attachment-chat-render, since it no
  longer pins anything.
